Update secure-app-model.md
The 'Using the Refresh Token' section is incorrect with its parameter use. Ref https://docs.microsoft.com/en-us/powershell/module/partnercenter/connect-partnercenter?view=partnercenterps-3.0

When using the AccessToken parameter, the ApplicationId should not be used. They are part of different parameter sets and will error out. In the example provided, either the ApplicationId should be removed, or the RefreshToken parameter should be used together with the ApplicationId parameter.